+++ plugin_ide.ui.tests/src-lang/melnorme/lang/ide/ui/editor/text/LangAutoEditStrategyTest.java
@@ -11,15 +11,6 @@
 package melnorme.lang.ide.ui.editor.text;
 
 import static melnorme.utilbox.core.Assert.AssertNamespace.assertTrue;
-import melnorme.lang.ide.core.text.BlockHeuristicsScannner;
-import melnorme.lang.ide.core.text.SamplePartitionScanner;
-import melnorme.lang.ide.core.text.Scanner_BaseTest;
-import melnorme.lang.ide.ui.CodeFormatterConstants;
-import melnorme.lang.ide.ui.CodeFormatterConstants.IndentMode;
-import melnorme.lang.ide.ui.editor.text.LangAutoEditPreferenceConstants;
-import melnorme.lang.ide.ui.editor.text.LangAutoEditStrategy;
-import melnorme.lang.ide.ui.text.util.LangAutoEditUtils;
-import melnorme.utilbox.misc.MiscUtil;
 
 import org.eclipse.jface.text.BadLocationException;
 import org.eclipse.jface.text.Document;
@@ -28,6 +19,15 @@ import org.eclipse.jface.text.IDocument;
 import org.eclipse.jface.text.IRegion;
 import org.eclipse.swt.SWT;
 import org.junit.Test;
+
+import melnorme.lang.ide.core.text.BlockHeuristicsScannner;
+import melnorme.lang.ide.core.text.SamplePartitionScanner;
+import melnorme.lang.ide.core.text.Scanner_BaseTest;
+import melnorme.lang.ide.ui.CodeFormatterConstants;
+import melnorme.lang.ide.ui.CodeFormatterConstants.IndentMode;
+import melnorme.lang.ide.ui.LangAutoEditPreferenceConstants;
+import melnorme.lang.ide.ui.text.util.LangAutoEditUtils;
+import melnorme.utilbox.misc.MiscUtil;
 
 public class LangAutoEditStrategyTest extends Scanner_BaseTest {