+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/editor/text/LangAutoEditStrategyExt.java
@@ -14,6 +14,7 @@ package melnorme.lang.ide.ui.editor.text;
 import static melnorme.utilbox.core.CoreUtil.array;
 import melnorme.lang.ide.core.text.BlockHeuristicsScannner;
 import melnorme.lang.ide.core.text.BlockHeuristicsScannner.BlockTokenRule;
+import melnorme.lang.ide.ui.LangAutoEditPreferenceConstants;
 import melnorme.utilbox.misc.ArrayUtil;
 
 import org.eclipse.jface.text.BadLocationException;