+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/preferences/LangEditorConfigurationBlock.java
@@ -8,17 +8,20 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation (JDT)
  *     DLTK team ? - DLTK modifications 
- *     Bruno Medeiros - Lang modifications
+ *     Bruno Medeiros - Lang rewrite
  *******************************************************************************/
 package melnorme.lang.ide.ui.preferences;
 
+import melnorme.lang.ide.ui.ContentAssistConstants;
+import melnorme.lang.ide.ui.EditorSettings_Actual.EditorPrefConstants;
 import melnorme.lang.ide.ui.preferences.EditorAppearanceColorsComponent.EditorColorItem;
 import melnorme.lang.ide.ui.preferences.common.AbstractComponentsPrefPage;
 import melnorme.lang.ide.ui.preferences.common.AbstractPreferencesBlock;
+import melnorme.util.swt.components.fields.CheckBoxField;
 
+import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Composite;
 
-//originally from DLTK version 5.0.0
 public class LangEditorConfigurationBlock extends AbstractPreferencesBlock {
 	
 	public LangEditorConfigurationBlock(AbstractComponentsPrefPage mainPreferencePage) {
@@ -36,6 +39,10 @@ public class LangEditorConfigurationBlock extends AbstractPreferencesBlock {
 	}
 	
 	protected void createAppearanceGroup(Composite parent) {
+		createBooleanField(parent,
+			EditorPrefConstants.MATCHING_BRACKETS,
+			new CheckBoxField(PreferencesMessages.EditorPreferencePage_matchingBrackets));
+		
 		createEditorAppearanceColorsComponent(parent);
 	}
 	
@@ -50,7 +57,22 @@ public class LangEditorConfigurationBlock extends AbstractPreferencesBlock {
 	}
 	
 	protected EditorColorItem[] createEditorAppearanceColorEntries() {
-		return new EditorColorItem[] { };
+		return new EditorColorItem[] { 
+			new EditorColorItem(
+				PreferencesMessages.EditorPreferencePage_matchingBracketsHighlightColor,
+				EditorPrefConstants.MATCHING_BRACKETS_COLOR),
+			new EditorColorItem(
+				PreferencesMessages.EditorPreferencePage_backgroundForMethodParameters,
+				ContentAssistConstants.PARAMETERS_BACKGROUND),
+			new EditorColorItem(
+				PreferencesMessages.EditorPreferencePage_foregroundForMethodParameters,
+				ContentAssistConstants.PARAMETERS_FOREGROUND),
+			new EditorColorItem(
+				PreferencesMessages.EditorPreferencePage_sourceHoverBackgroundColor,
+				EditorPrefConstants.SOURCE_HOVER_BACKGROUND_COLOR_rgb.key,
+				EditorPrefConstants.SOURCE_HOVER_BACKGROUND_COLOR_UseSystemDefault.key,
+				SWT.COLOR_INFO_BACKGROUND),
+		};
 	}
 	
 }