+++ plugin_ide.ui/src/melnorme/lang/ide/ui/CodeFormatterConstants_Actual.java
@@ -20,15 +20,4 @@ public interface CodeFormatterConstants_Actual {
 	IntPreference FORMATTER_INDENTATION_SPACES_SIZE = 
 			new IntPreference(QUALIFIER, "CodeFormatterConstants.FORMATTER_INDENTATION_SIZE", 4);
 	
-	class Helper {
-		
-		public static void initDefaults() {
-			// This will ensure default value is initialized
-			FORMATTER_INDENT_MODE.get();
-			FORMATTER_TAB_SIZE.get();
-			FORMATTER_INDENTATION_SPACES_SIZE.get();
-		}
-		
-	}
-	
 }