+++ plugin_ide.ui/src/melnorme/lang/ide/ui/EditorPreferences.java
@@ -1,6 +1,0 @@
-package melnorme.lang.ide.ui;
-
-
-public interface EditorPreferences extends EditorPreferences_Default {
-	
-}