+++ plugin_ide.ui/src/melnorme/lang/ide/ui/LangAutoEditPreferenceConstants.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2010, 2011 IBM Corporation and others.
+ * Copyright (c) 2015 Bruno Medeiros and other Contributors.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,11 +8,11 @@
  * Contributors:
  *     Bruno Medeiros - initial API and implementation
  *******************************************************************************/
-package melnorme.lang.ide.ui.editor.text;
+package melnorme.lang.ide.ui;
 
-import melnorme.lang.ide.ui.LangAutoEditPreferenceConstants_Actual;
+import melnorme.lang.tooling.LANG_SPECIFIC;
 
-
-public interface LangAutoEditPreferenceConstants extends LangAutoEditPreferenceConstants_Actual {
+@LANG_SPECIFIC
+public interface LangAutoEditPreferenceConstants extends LangAutoEditPreferenceConstants_Default {
 	
 }